+++ DX11Renderer/DX11Renderer/BaseBufferData.h
@@ -6,7 +6,7 @@ class BaseBufferData
 {
 public:
 	virtual D3D11_SUBRESOURCE_DATA GetSubresourceData() const = 0;
-	virtual size_t GetElementCount() const = 0;
-	virtual size_t GetSizeInBytes() const = 0;
-	virtual size_t GetStride() const = 0;
+	virtual UINT GetElementCount() const = 0;
+	virtual UINT GetSizeInBytes() const = 0;
+	virtual UINT GetStride() const = 0;
 };
